requirements.txt
playwright>=1.44.0
requests>=2.33.0
opencv-python>=4.9.0
Pillow>=10.3.0
pandas>=2.2.0
python-dotenv>=1.0.0
crewai>=0.55.0
crewai-tools>=0.8.0
numpy>=1.26.0
anthropic>=0.28.0
openai>=1.30.0
python-dateutil>=2.9.0
aiofiles>=23.2.1
